% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKld.R
\name{computeKld}
\alias{computeKld}
\title{Compute KLD}
\usage{
computeKld(x, pseudocount = 1e-08)
}
\arguments{
\item{x}{matrix-like object or vector with non-negative numbers}

\item{pseudocount}{pseudocount = this value is added to each zero}
}
\value{
kullback-leibler dissimilarity matrix
}
\description{
Compute Kullback-Leibler dissimilarity
}
\details{
Outputs a Kullback-Leibler dissimilarity (symmetrized divergence) matrix.Note:
}
