% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPval.R
\name{getPval}
\alias{getPval}
\title{Get p-value}
\usage{
getPval(matrix, x.index, y.index, N.rand = 1000, method = "spearman",
  renorm = F, permutandboot = F, plot = F, verbose = F)
}
\arguments{
\item{matrix}{input matrix}

\item{x.index}{index of first vector in the input matrix}

\item{y.index}{index of second vector in the input matrix}

\item{N.rand}{number of iterations used for the permutation test}

\item{method}{similarity measure (supported measures are: "pearson", "spearman", "bray" and "kld")}

\item{renorm}{renormalize after permutation}

\item{permutandboot}{compute a bootstrap distribution in addition to the permutation distribution and}

\item{plot}{plot the histogram of the permutation and, if permutandboot is true, of the bootstrap distribution}

\item{verbose}{print distribution properties and p-value}
}
\value{
p-value of the association
}
\description{
Get permuation-based p-value for association between two vectors.
}
\details{
# Compute the association between two vectors using the given method and
compute its p-value using a permutation test. This method was adapted from R code by Fah Sahtirapongsasuti.
This method was adapted from CCREPE: http://huttenhower.sph.harvard.edu/ccrepe.
Emma Schwager et al Detecting statistically significant associtations between sparse and high dimensional compositional data. (In progress).
}
